refactor(example): use barrel import for core module (#1869)
