+++ projects/example-app/src/app/core/containers/index.ts
@@ -1,0 +1,2 @@
+export * from './app.component';
+export * from './not-found-page.component';
