+++ projects/example-app/src/app/core/index.ts
@@ -1,0 +1,1 @@
+export * from './core.module';
